Refactor: 2순위 장르 수정 API 로직 변경 (#21)
