+++ src/main/java/tavebalak/OTTify/genre/repository/UserGenreRepository.java
@@ -21,6 +21,9 @@ public interface UserGenreRepository extends JpaRepository<UserGenre, Long> {
     @Query("select ug from UserGenre ug where ug.user.id =:userId and ug.isFirst = false")
     List<UserGenre> find2ndGenreByUserId(@Param("userId") Long userId);
 
+    Boolean existsByGenreIdAndUserIdAndIsFirst(@Param("genreId") Long genreId, @Param("userId") Long userId, boolean isFirst);
+    UserGenre findByGenreIdAndUserIdAndIsFirst(@Param("genreId") Long genreId, @Param("userId") Long userId, boolean isFirst);
+
     @Transactional
     @Modifying
     @Query("delete from UserGenre ug where ug.genre.id in :genreIds and ug.user.id =:userId")
