+++ src/main/java/tavebalak/OTTify/user/controller/UserController.java
@@ -6,11 +6,9 @@ import org.springframework.http.HttpStatus;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import tavebalak.OTTify.common.BaseResponse;
-import tavebalak.OTTify.genre.dto.FirstGenreUpdateRequestDTO;
-import tavebalak.OTTify.genre.dto.SecondGenreUpdateRequestDTO;
+import tavebalak.OTTify.genre.dto.request.GenreUpdateDTO;
+import tavebalak.OTTify.genre.dto.request.SecondGenreUpdateRequestDTO;
 import tavebalak.OTTify.user.service.UserServiceImpl;
-
-import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
@@ -22,13 +20,13 @@ public class UserController {
 
     @PatchMapping("/{id}/1stGenre")
     @ResponseStatus(HttpStatus.OK)
-    public BaseResponse<Long> update1stLikedGenre(@PathVariable("id") Long userId, @Validated @RequestBody FirstGenreUpdateRequestDTO updateRequestDto) {
+    public BaseResponse<Long> update1stLikedGenre(@PathVariable("id") Long userId, @Validated @RequestBody GenreUpdateDTO updateRequestDto) {
         return BaseResponse.success(userService.update1stGenre(userId, updateRequestDto));
     }
 
     @PatchMapping("/{id}/2ndGenre")
     @ResponseStatus(HttpStatus.OK)
-    public BaseResponse<Long> update2ndLikedGenre(@PathVariable("id") Long userId, @Validated @RequestBody List<SecondGenreUpdateRequestDTO> updateRequestDTO) {
+    public BaseResponse<Long> update2ndLikedGenre(@PathVariable("id") Long userId, @Validated @RequestBody SecondGenreUpdateRequestDTO updateRequestDTO) {
         return BaseResponse.success(userService.update2ndGenre(userId, updateRequestDTO));
     }
 }
