+++ src/main/java/tavebalak/OTTify/user/service/UserService.java
@@ -1,12 +1,10 @@
 package tavebalak.OTTify.user.service;
 
-import tavebalak.OTTify.genre.dto.FirstGenreUpdateRequestDTO;
-import tavebalak.OTTify.genre.dto.SecondGenreUpdateRequestDTO;
-
-import java.util.List;
+import tavebalak.OTTify.genre.dto.request.GenreUpdateDTO;
+import tavebalak.OTTify.genre.dto.request.SecondGenreUpdateRequestDTO;
 
 public interface UserService {
 
-    Long update1stGenre(Long userId, FirstGenreUpdateRequestDTO updateRequestDTO);
-    Long update2ndGenre(Long userId, List<SecondGenreUpdateRequestDTO> updateRequestDTO);
+    Long update1stGenre(Long userId, GenreUpdateDTO updateRequestDTO);
+    Long update2ndGenre(Long userId, SecondGenreUpdateRequestDTO updateRequestDTO);
 }
