+++ src/main/java/tavebalak/OTTify/user/service/UserServiceImpl.java
@@ -6,17 +6,13 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import tavebalak.OTTify.error.ErrorCode;
 import tavebalak.OTTify.error.exception.NotFoundException;
-import tavebalak.OTTify.genre.dto.FirstGenreUpdateRequestDTO;
-import tavebalak.OTTify.genre.dto.SecondGenreUpdateRequestDTO;
+import tavebalak.OTTify.genre.dto.request.GenreUpdateDTO;
+import tavebalak.OTTify.genre.dto.request.SecondGenreUpdateRequestDTO;
 import tavebalak.OTTify.genre.entity.Genre;
 import tavebalak.OTTify.genre.entity.UserGenre;
 import tavebalak.OTTify.genre.repository.GenreRepository;
 import tavebalak.OTTify.genre.repository.UserGenreRepository;
-import tavebalak.OTTify.user.entity.User;
 import tavebalak.OTTify.user.repository.UserRepository;
-
-import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 @Transactional(readOnly = true)
@@ -30,7 +26,7 @@ public class UserServiceImpl implements UserService {
 
     @Override
     @Transactional
-    public Long update1stGenre(Long userId, FirstGenreUpdateRequestDTO updateRequestDTO) {
+    public Long update1stGenre(Long userId, GenreUpdateDTO updateRequestDTO) {
         UserGenre userGenre = userGenreRepository.findByUserIdAndIsFirst(userId, true)
                 .orElseThrow(() -> new NotFoundException(ErrorCode.ENTITY_NOT_FOUND));
         Genre genre = genreRepository.findById(updateRequestDTO.getId())
@@ -43,54 +39,27 @@ public class UserServiceImpl implements UserService {
 
     @Override
     @Transactional
-    public Long update2ndGenre(Long userId, List<SecondGenreUpdateRequestDTO> updateRequestDTO) {
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new NotFoundException(ErrorCode.ENTITY_NOT_FOUND));
+    public Long update2ndGenre(Long userId, SecondGenreUpdateRequestDTO updateRequestDTO) {
+        // req로 들어온 id 값이 유효한 장르 id인지 확인
+        Genre genre = genreRepository.findById(updateRequestDTO.getId())
+                .orElseThrow(() -> new NotFoundException(ErrorCode.GENRE_NOT_FOUND));
 
-        // 이전에 선택한 2순위 장르 리스트
-        List<Long> preGenreList = userGenreRepository.find2ndGenreByUserId(userId).stream()
-                .map((UserGenre ug) -> ug.getGenre().getId())
-                .collect(Collectors.toList());
+        // genreId와 userId로 UserGenre 조회
+        boolean isUserGenreExists = userGenreRepository.existsByGenreIdAndUserIdAndIsFirst(genre.getId(), userId, false);
 
-        // 현재 선택한 2순위 장르 리스트
-        List<Long> nowGenreList = updateRequestDTO.stream()
-                .map(SecondGenreUpdateRequestDTO::getId)
-                .collect(Collectors.toList());
-
-        if (!preGenreList.isEmpty()) { // 이전 선택한 2순위 장르가 있는 경우
-            // 삭제할 장르 리스트 - 이전 리스트에는 있는데 현재 리스트에는 없는 장르
-            List<Long> deleteGenres = preGenreList.stream()
-                    .filter(g -> !nowGenreList.contains(g))
-                    .collect(Collectors.toList());
-            userGenreRepository.deleteAllByIdInQuery(deleteGenres, userId);
-
-            // 추가할 장르 리스트 - 이전 리스트에는 없는데 현재 리스트에는 있는 장르
-            List<Long> insertGenres = nowGenreList.stream()
-                    .filter(g -> !preGenreList.contains(g))
-                    .collect(Collectors.toList());
-            insertGenres.stream()
-                    .forEach(g -> {
-                        UserGenre userGenre = UserGenre.create(
-                                user,
-                                genreRepository.findById(g)
-                                        .orElseThrow(() -> new NotFoundException(ErrorCode.GENRE_NOT_FOUND)),
-                                false
-                        );
-                        userGenreRepository.save(userGenre);
-                    });
-        } else { // 이전 선택한 2순위 장르가 없는 경우
-            nowGenreList.stream()
-                    .forEach(g -> {
-                        UserGenre userGenre = UserGenre.create(
-                                user,
-                                genreRepository.findById(g)
-                                        .orElseThrow(() -> new NotFoundException(ErrorCode.GENRE_NOT_FOUND)),
-                                false
-                        );
-                        userGenreRepository.save(userGenre);
-                    });
+        // 조회된 UserGenre가 없을 경우 저장 & 있을 경우 삭제
+        if (isUserGenreExists) {
+            UserGenre findUserGenre = userGenreRepository.findByGenreIdAndUserIdAndIsFirst(genre.getId(), userId, false);
+            userGenreRepository.delete(findUserGenre);
+        } else {
+            userGenreRepository.save(UserGenre.builder()
+                    .genre(genre)
+                    .user(userRepository.findById(userId)
+                            .orElseThrow(() -> new NotFoundException(ErrorCode.ENTITY_NOT_FOUND)))
+                    .build());
         }
 
         return userId;
     }
+
 }
